Move licensing resolver into enterprise/cmd/frontend
Not needed by other services, so it should live here.
